restructure api into version
