+++ app.js
@@ -7,7 +7,7 @@ const app = express();
 const path = require('path');
 import {router as apiRoot} from "./routes/api-root";
 import {decode_auth} from "./lib/decode_auth";
-import {authStrategy} from "./routes/auth/strategy";
+import {authStrategy} from "./routes/v1/auth/strategy";
 const passport = require('passport');
 
 passport.use(authStrategy);
