+++ routes/api-root.js
@@ -1,7 +1,8 @@
 import express from 'express';
 const router = express.Router();
-import {authRouter} from "./auth/auth";
-import {surveyImportRouter} from "./survey_import/survey_import";
+import {router as v1} from './v1/v1';
+import {authRouter} from "./v1/auth/auth";
+import {surveyImportRouter} from "./v1/survey_import/survey_import";
 const passport = require('passport');
 
 const auth = passport.authenticate('jwt', {session: false});
@@ -14,8 +15,9 @@ router.get('/', (req, res) => {
   res.json(message);
 });
 
-router.use('/auth', authRouter);
-router.use('/survey_import', auth, surveyImportRouter);
+router.use('/v1', v1);
+// router.use('/auth', authRouter);
+// router.use('/survey_import', auth, surveyImportRouter);
 
 
 export {router}
