+++ routes/v1/auth/strategy.js
