+++ routes/v1/v1.js
@@ -1,0 +1,13 @@
+import express from 'express';
+import
+const router = express.Router();
+
+router.get('/', (req, res) => {
+  res.json({
+    message: 'api v1 root'
+  });
+});
+
+
+
+export {router};
